Add files via upload
We were asked to predict the price of a piece of diamond. Which sometimes can be a tricky thing.

Indistinguishable:
Diamonds can look the same at the first look. But it’s the  quality that defines the price

Quality???
It’s quite hard to compare two pieces of diamonds. There is no particular definition for “quality” here.

This Presentation will show you how we went through the details of predicting the best price for a diamond piece!